requirements-bot.txt
aiogram>=3.4,<4
SQLAlchemy[asyncio]>=2.0
aiosqlite>=0.19
python-dotenv>=1.0
openai>=1.40
Pillow>=10.0
pytest>=8.0
pytest-asyncio>=0.24